+++ app/(tabs)/folder/_components/serachSection/SearchSection.tsx
